Kraken - Avoid fetching full ledger for funding
When getFundingHistory(...) is called, KrakenAccountService first
downloads the full ledger, which contains trades, withdrawal and
deposit events, and then filters out trade events. This causes a large
number of API calls to be made, while only a small fraction of the
ledger is used. The Kraken API offers the possibility to retrieve only
some types of events (e.g., trade, deposit or withdrawal events). Only
download deposit and withdrawal events when getFundingHistory(...) is
called.