Floats don't have enough precision to store timestamps.
When using xchange-stream's Bitfinex implementation it fakes the timestamp because Bitfinex somehow forgot to include one in their Ticker payload. It runs `System.currentTimeMillis() / 1000` which, when stored as a float, was giving me times that were sometimes as much as 90 seconds in the future. Fixing the BitfinexTicker to use a double instead has made the math work properly and my timestamps are now all in the past, where they're supposed to be.