[abucoins] API was updated on abucoins, adding new methods
Also replacing deprecated calls with current naming convetion.

Also, ran into a situation where a value that was modeled as an enum
was returned with a new value that I hadn't seen in the documentation.
This causes the mapping to fail during the json marshalling.

To improve the failure behavior in this case, I internally switched the
field to be a String so that json parsing would succeed.  I do however
continue to use the enum at the Pojo API level, it just treats strings
it doesn't recognized as 'unknown'.  This seems like more desirable
behavior as the API continues to work, it may just be that your status
isn't mapped to a known value.  If you aren't using the status, has no
impact, if you need the status, then there's a raw method for obtaining
the string value.  It made debugging much easier.