bugfix/ETD-1130 - added LedgerType SALE ( tx - spend and received.
https://support.kraken.com/hc/en-us/articles/360001169383-How-to-interpret-Ledger-history-fields

"sale" = this shows only as a filter option when viewing your Ledger history while signed in to your account, it is not shown in history exports. "Sale" is a filter that brings up all "Spend" and "Receive" entries from orders placed via the new Kraken app or Buy Crypto button.